requirements.txt
tensorflow>=2.15.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.8.0
streamlit>=1.29.0